backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
redis==5.0.1
openai==1.12.0
playwright==1.42.0
faiss-cpu==1.7.4
numpy==1.26.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
websockets==12.0
structlog==24.1.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.1.0
